Refactor semantic analysis to break circular dependencies using Trait Injection
