+++ src/semantic/control_flow.rs
@@ -8,19 +8,21 @@
 //! - Break/Continue (παῦε, συνέχιζε)
 
 use super::expressions::get_first_word;
+use super::conversion::convert_assembled_to_analyzed;
 use super::{
-    AnalyzedExpr, AnalyzedExprKind, AnalyzedStatement, GlossaType, Scope, analyze_statement,
-    assemble_statement, convert_assembled_to_analyzed,
+    AnalyzedExpr, AnalyzedExprKind, AnalyzedStatement, GlossaType, Scope, assemble_statement,
 };
 use crate::ast::{Clause, Expr, Statement};
 use crate::errors::GlossaError;
 use crate::morphology::lexicon;
+use crate::semantic::traits::StatementAnalyzer;
 
 /// Check if a statement is a control flow construct and analyze it
 /// Returns Some(AnalyzedStatement) if it's control flow, None otherwise
 pub fn analyze_control_flow(
     stmt: &Statement,
     scope: &mut Scope,
+    analyzer: &mut impl StatementAnalyzer,
 ) -> Result<Option<AnalyzedStatement>, GlossaError> {
     // Note: Function definitions are handled in `mod.rs` before calling this.
 
@@ -33,22 +35,22 @@ pub fn analyze_control_flow(
 
     // Conditional: εἰ/ἐάν condition, body [, εἰ δὲ μή, else_body]
     if lexicon::is_conditional_particle(&normalized) {
-        return parse_conditional(stmt, scope, 0);
+        return parse_conditional(stmt, scope, analyzer, 0);
     }
 
     // While loop: ἕως condition, body
     if normalized == "εως" {
-        return parse_while_loop(stmt, scope);
+        return parse_while_loop(stmt, scope, analyzer);
     }
 
     // For loop with range: ἀπὸ start μέχρι/ἕως end, body
     if normalized == "απο" {
-        return parse_for_range_loop(stmt, scope);
+        return parse_for_range_loop(stmt, scope, analyzer);
     }
 
     // For loop with iteration: διὰ collection, body
     if normalized == "δια" {
-        return parse_for_iteration_loop(stmt, scope);
+        return parse_for_iteration_loop(stmt, scope, analyzer);
     }
 
     if lexicon::is_loop_particle(&normalized) {
@@ -57,7 +59,7 @@ pub fn analyze_control_flow(
     }
 
     if lexicon::is_match_particle(&normalized) {
-        return parse_match_expression(stmt, scope);
+        return parse_match_expression(stmt, scope, analyzer);
     }
 
     // Return: δός (give)
@@ -84,6 +86,7 @@ pub fn analyze_control_flow(
 fn parse_while_loop(
     stmt: &Statement,
     scope: &mut Scope,
+    analyzer: &mut impl StatementAnalyzer,
 ) -> Result<Option<AnalyzedStatement>, GlossaError> {
     if stmt.clauses().len() < 2 {
         return Err(GlossaError::semantic(
@@ -105,7 +108,7 @@ fn parse_while_loop(
     };
 
     // Analyze body using unified helper
-    let body_analyzed = analyze_statement(&body_stmt, scope)?;
+    let body_analyzed = analyzer.analyze(&body_stmt, scope)?;
 
     Ok(Some(AnalyzedStatement::While {
         condition: Box::new(condition_expr),
@@ -118,6 +121,7 @@ fn parse_while_loop(
 fn parse_for_range_loop(
     stmt: &Statement,
     scope: &mut Scope,
+    analyzer: &mut impl StatementAnalyzer,
 ) -> Result<Option<AnalyzedStatement>, GlossaError> {
     if stmt.clauses().len() < 2 {
         return Err(GlossaError::semantic(
@@ -242,7 +246,7 @@ fn parse_for_range_loop(
         let mut loop_scope = scope.enter_scope();
         loop_scope.define(variable.clone(), GlossaType::Number);
 
-        analyze_statement(&body_stmt, &mut loop_scope)?
+        analyzer.analyze(&body_stmt, &mut loop_scope)?
     };
 
     Ok(Some(AnalyzedStatement::For {
@@ -258,6 +262,7 @@ fn parse_for_range_loop(
 fn parse_for_iteration_loop(
     stmt: &Statement,
     scope: &mut Scope,
+    analyzer: &mut impl StatementAnalyzer,
 ) -> Result<Option<AnalyzedStatement>, GlossaError> {
     if stmt.clauses().len() < 2 {
         return Err(GlossaError::semantic(
@@ -328,7 +333,7 @@ fn parse_for_iteration_loop(
         // TODO: Infer element type from collection type
         loop_scope.define(variable.clone(), GlossaType::String);
 
-        analyze_statement(&body_stmt, &mut loop_scope)?
+        analyzer.analyze(&body_stmt, &mut loop_scope)?
     };
 
     Ok(Some(AnalyzedStatement::For {
@@ -344,6 +349,7 @@ fn parse_for_iteration_loop(
 fn parse_match_expression(
     stmt: &Statement,
     scope: &mut Scope,
+    analyzer: &mut impl StatementAnalyzer,
 ) -> Result<Option<AnalyzedStatement>, GlossaError> {
     if stmt.clauses().is_empty() {
         return Err(GlossaError::semantic(
@@ -391,7 +397,7 @@ fn parse_match_expression(
                 is_query: false,
                 is_propagate: false,
             };
-            let body_analyzed = analyze_statement(&body_stmt, scope)?;
+            let body_analyzed = analyzer.analyze(&body_stmt, scope)?;
 
             arms.push((pattern, body_analyzed));
         }
@@ -572,6 +578,7 @@ const MAX_CONTROL_FLOW_DEPTH: usize = 100;
 fn parse_conditional(
     stmt: &Statement,
     scope: &mut Scope,
+    analyzer: &mut impl StatementAnalyzer,
     depth: usize,
 ) -> Result<Option<AnalyzedStatement>, GlossaError> {
     if depth > MAX_CONTROL_FLOW_DEPTH {
@@ -605,7 +612,7 @@ fn parse_conditional(
             is_query: false,
             is_propagate: false,
         };
-        analyze_statement(&stmt, scope)?
+        analyzer.analyze(&stmt, scope)?
     };
 
     // Check for else/elif clause
@@ -619,7 +626,7 @@ fn parse_conditional(
                 is_query: false,
                 is_propagate: false,
             };
-            Some(analyze_statement(&stmt, scope)?)
+            Some(analyzer.analyze(&stmt, scope)?)
         }
         // Check if it's "εἰ" or "ἐάν" (elif)
         else if check_conditional_start(second_expr) {
@@ -641,7 +648,7 @@ fn parse_conditional(
             };
 
             // Recursively parse as a new conditional (which becomes the else body)
-            if let Some(elif_analyzed) = parse_conditional(&elif_stmt, scope, depth + 1)? {
+            if let Some(elif_analyzed) = parse_conditional(&elif_stmt, scope, analyzer, depth + 1)? {
                 Some(vec![elif_analyzed])
             } else {
                 None
